imagegenerator/requirements.txt
streamlit==1.15.0
requests==2.26.0
Pillow==8.3.2
altair==4.1.0